Controller Chnages

+++ src/main/java/com/ubi/academicapplication/service/StudentServiceImpl.java
@@ -3,11 +3,8 @@ package com.ubi.academicapplication.service;
 import java.util.List;
 import java.util.Optional;
 
-import com.ubi.academicapplication.entity.Payment;
-import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 
 import com.ubi.academicapplication.dto.responsedto.Response;
 import com.ubi.academicapplication.entity.Student;
@@ -15,8 +12,6 @@ import com.ubi.academicapplication.error.CustomException;
 import com.ubi.academicapplication.error.HttpStatusCode;
 import com.ubi.academicapplication.error.Result;
 import com.ubi.academicapplication.repository.StudentRepository;
-
-
 
 @Service
 public class StudentServiceImpl implements StudentService {
@@ -27,19 +22,20 @@ public class StudentServiceImpl implements StudentService {
 	@Autowired
 	private StudentRepository repository;
 
-	public Student saveStudent(Student student) {
+	public Response<Student> saveStudent(Student student) {
+
+		Response<Student> response = new Response<>();
+
 		if (student.getStudentName().isEmpty() || student.getStudentName().length() == 0) {
 			throw new CustomException(HttpStatusCode.NO_STUDENT_NAME_FOUND.getCode(),
 					HttpStatusCode.NO_STUDENT_NAME_FOUND, HttpStatusCode.NO_STUDENT_NAME_FOUND.getMessage(), res);
-		}Student savedStudent = repository.save(student);
-		return savedStudent;
+		}
+		Student savedStudent = repository.save(student);
+		response.setStatusCode(HttpStatusCode.RESOURCE_CREATED_SUCCESSFULLY.getCode());
+		response.setMessage(HttpStatusCode.RESOURCE_CREATED_SUCCESSFULLY.getMessage());
+		response.setResult(new Result<Student>(savedStudent));
+		return response;
 	}
-
-//	public Response<List<Student>> getStudents(Integer PageNumber, Integer PageSize) {
-//		Pageable paging = PageRequest.of(PageNumber, PageSize);
-//		Page<Student> pagedResult = repository.findAll(paging);
-//		return ;
-//	}
 
 	public Response<List<Student>> getStudents() {
 		Response<List<Student>> getListofStudent = new Response<List<Student>>();
@@ -72,46 +68,49 @@ public class StudentServiceImpl implements StudentService {
 	}
 
 	@Override
-	public void deleteStudent(int id) {
-		try {
-			Response<Student> deleteStd=new Response<Student>();
-			repository.deleteById(id);
-			
-		} catch (IllegalArgumentException e) {
-			throw new CustomException(HttpStatusCode.NO_STUDENT_FOUND.getCode(), HttpStatusCode.NO_STUDENT_FOUND,
-					HttpStatusCode.NO_STUDENT_FOUND.getMessage(), res);
-		} catch (Exception e) {
-			throw new CustomException(HttpStatusCode.NO_STUDENT_FOUND.getCode(), HttpStatusCode.NO_STUDENT_FOUND,
-					HttpStatusCode.NO_STUDENT_FOUND.getMessage(), res);
-
+	public Response<Student> deleteById(int id) {
+		res.setData(null);
+		Optional<Student> student = repository.findById(id);
+		if (!student.isPresent()) {
+			throw new CustomException(HttpStatusCode.RESOURCE_NOT_FOUND.getCode(), HttpStatusCode.RESOURCE_NOT_FOUND,
+					HttpStatusCode.RESOURCE_NOT_FOUND.getMessage(), res);
 		}
+		repository.deleteById(id);
+		Response<Student> response = new Response<>();
+		response.setMessage(HttpStatusCode.STUDENT_DELETED.getMessage());
+		response.setStatusCode(HttpStatusCode.STUDENT_DELETED.getCode());
+		response.setResult(new Result<Student>(student.get()));
+		return response;
 	}
 
-	public Student updateStudent(Student student) {	
+	public Response<Student> updateStudent(Student student) {
 		res.setData(null);
 		Optional<Student> existingStudentContainer = repository.findById(student.getStudentId());
-		if (!existingStudentContainer.isPresent())
+		if (!existingStudentContainer.isPresent()) {
 			throw new CustomException(HttpStatusCode.NO_STUDENT_FOUND.getCode(), HttpStatusCode.NO_STUDENT_FOUND,
 					HttpStatusCode.NO_STUDENT_FOUND.getMessage(), res);
-		else {
-			Student existingStudent = existingStudentContainer.get();
-			existingStudent.setStudentName(student.getStudentName());
-			existingStudent.setStudentStatus(student.isStudentStatus());
-			existingStudent.setCategory(student.getCategory());
-			existingStudent.setFatherName(student.getFatherName());
-			existingStudent.setFatherOccupation(student.getFatherOccupation());
-			existingStudent.setMotherName(student.getMotherName());
-			existingStudent.setMotherOccupation(student.getMotherOccupation());
-			existingStudent.setGender(student.getGender());
-			existingStudent.setJoiningDate(student.getJoiningDate());
-			existingStudent.setStatus(student.getStatus());
-			existingStudent.setVerifiedByTeacher(student.getVerifiedByTeacher());
-			existingStudent.setVerifiedByPrincipal(student.getVerifiedByPrincipal());
-			existingStudent.setVerifiedByRegion(student.getVerifiedByRegion());
-			return repository.save(existingStudent);
 		}
+		Student existingStudent = existingStudentContainer.get();
+		existingStudent.setStudentName(student.getStudentName());
+		existingStudent.setStudentStatus(student.isStudentStatus());
+		existingStudent.setCategory(student.getCategory());
+		existingStudent.setFatherName(student.getFatherName());
+		existingStudent.setFatherOccupation(student.getFatherOccupation());
+		existingStudent.setMotherName(student.getMotherName());
+		existingStudent.setMotherOccupation(student.getMotherOccupation());
+		existingStudent.setGender(student.getGender());
+		existingStudent.setJoiningDate(student.getJoiningDate());
+		existingStudent.setStatus(student.getStatus());
+		existingStudent.setVerifiedByTeacher(student.getVerifiedByTeacher());
+		existingStudent.setVerifiedByPrincipal(student.getVerifiedByPrincipal());
+		existingStudent.setVerifiedByRegion(student.getVerifiedByRegion());
 
+		Student updateStudent = repository.save(existingStudent);
+		Response<Student> response = new Response<>();
+		response.setMessage(HttpStatusCode.STUDENT_UPDATED.getMessage());
+		response.setStatusCode(HttpStatusCode.STUDENT_UPDATED.getCode());
+		response.setResult(new Result<>(updateStudent));
+		return response;
 	}
-
 
 }
